zerolight

+++ src/main/java/com/daepamarket/daepa_market_backend/admin/review/AllReviewDTO.java
@@ -2,18 +2,19 @@ package com.daepamarket.daepa_market_backend.admin.review;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
 import java.time.LocalDateTime;
 
 @Data
 @AllArgsConstructor
 public class AllReviewDTO {
-    private String id;  // Long → String으로 변경
-    private Long realId;
+    private String id;          // "S-1" / "B-10"
+    private Long realId;        // 실제 PK
     private String product;
     private String buyer;
     private String seller;
-    private Long rating;
+    private Integer rating;     // ★ 여기 Long → Integer 로
     private String comment;
     private LocalDateTime date;
-    private String type; // BUY / SELL
+    private String type;        // "BUYER" / "SELLER"
 }
